renaming postalcode-en-* to locale-en-*
