+++ test/all.js
@@ -1,4 +1,4 @@
 require('./parse');
-require('./postalcode-en-AU');
-require('./postalcode-en-US');
+require('./locale-en-AU');
+require('./locale-en-US');
 require('./street-types');
